fix: Do not consider strings as iterable when preparing query parameters
When preparing query parameters for the API requests, parameters
received as strings should not be modified.

Before this change, if a string was passed as a parameter, the code
would override it with a version where each character was separated by
a comma. (e.g. `'string' -> 's,t,r,i,n,g'`)